+++ src/corpus/views/CorpusExportView.vue
@@ -1,10 +1,14 @@
 <script setup lang="ts">
+import { computed } from "vue";
+import { useRoute } from "vue-router";
 import ExportViewTemplate from "@/exports/ExportViewTemplate.vue";
+import useResourceIdParam from "@/resource/resourceIdParam.composable";
 
-defineProps<{
-  id: string;
-  path: string;
-}>();
+const route = useRoute();
+const id = useResourceIdParam();
+
+/** The export file path as read from the 'path' query param */
+const path = computed(() => String(route.query["path"]));
 </script>
 
 <template>